feat: validate only account with `type === 'user'` can login
