+++ src/shared/models/userInterfaces.ts
@@ -5,21 +5,6 @@ import {
   IVendorDetail,
 } from './generalInterfaces';
 
-export interface ILoginResponseRoot {
-  user_id: string;
-  email: string;
-  token: string;
-  permissions: string[];
-  type: string;
-  role_id: number;
-  role_name: string;
-}
-
-export interface ILoginPayloadRoot {
-  email: string;
-  password: string;
-}
-
 export interface IAllUserResponse {
   id: string;
   name: string;
